+++ cernml/coi/_configurable.py
@@ -167,6 +167,38 @@ class Config:
         """Return a read-only view of all declared fields."""
         return self._fields.values()
 
+    def get_field_values(self) -> t.Dict[str, t.Any]:
+        """Return a :class:`dict` of the pre-configured field values.
+
+        Note that this is not quite the expected input to
+        :meth:`validate_all()`; the latter expects the dict values to be
+        strings that convert cleanly back to the original field type.
+        This is the case for strings and numbers, but not e.g. for
+        boolean values.
+
+        Example:
+
+            >>> config = Config().add("flag", False).add("count", 10)
+            >>> values = config.get_field_values()
+            >>> values
+            {'flag': False, 'count': 10}
+
+        Passing this dict to :meth:`validate_all()` accidentally works
+        even though the type signature doesn't match:
+
+            >>> config.validate_all(config.get_field_values())
+            namespace(count=10, flag=False)
+
+        Note that bool values don't allow you to convert everything to
+        strings blindly to make the types match:
+
+            >>> config.validate_all(
+            ...     {k: str(v) for k, v in values.items()}
+            ... )
+            namespace(count=10, flag=True)
+        """
+        return {dest: field.value for dest, field in self._fields.items()}
+
     def add(
         self,
         dest: str,